feat: added configs and fix database schema

+++ src/main/java/com/example/businessapp/controller/CustomerController.java
@@ -31,7 +31,7 @@ public class CustomerController {
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<Customer> getCustomerById(@PathVariable Long id) {
+    public ResponseEntity<Customer> getCustomerById(@PathVariable Integer id) {
         return handler.getCustomerById(id);
     }
 }
